feat: stream-b08-provider-ownership-cleanup-003 - wire production failure parsing through owned packages
Route live exit-failure classification through provider-owned packages, delete duplicated shared parser grammar, remove committed coverage artifacts, and keep Codex dispatch thin in shared orchestration.
