+++ pkg/workers/provider/claude/exitfailure/failure.go
@@ -1,17 +1,52 @@
-package provider
+package exitfailure
 
 import (
 	"bytes"
-	_ "embed"
 	"encoding/json"
 	"fmt"
 	"strconv"
 	"strings"
-	"sync"
 	"unicode/utf8"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
+
+type FailureInput struct {
+	Stdout   []byte
+	Stderr   []byte
+	ExitCode int
+}
+
+type FailureResult struct {
+	Reason  interfaces.WorkFailureType
+	Message string
+}
+
+const ThrottleFailureMessage = "Claude is temporarily unavailable due to rate or capacity limits."
+const TimeoutFailureMessage = "Claude request timed out."
+
+var sensitiveCommandEnvNameFragments = []string{
+	"TOKEN", "SECRET", "PASSWORD", "PASS", "KEY", "CREDENTIAL", "CREDENTIALS", "AUTH",
+	"ANTHROPIC", "OPENAI", "GEMINI", "GOOGLE_APPLICATION_CREDENTIALS",
+	"AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY",
+}
+
+func isRedactedCommandEnvKey(name string) bool {
+	normalized := strings.ToUpper(strings.TrimSpace(name))
+	if normalized == "" {
+		return false
+	}
+	for _, fragment := range sensitiveCommandEnvNameFragments {
+		if strings.Contains(normalized, fragment) {
+			return true
+		}
+	}
+	return false
+}
+
+func ParseProviderFailure(input FailureInput) FailureResult {
+	return parseProviderFailure(input)
+}
 
 const (
 	claudeFailureMessageBytes = 1024
@@ -38,7 +73,8 @@ type claudeStructuredFailure struct {
 // Structured scanning covers the complete captured streams while bounding each
 // candidate record; text fallback and actionable messages are separately
 // bounded so a failed command cannot publish its transcript.
-func ParseClaudeProviderFailure(result CommandResult) ProviderFailureResult {
+func parseProviderFailure(input FailureInput) FailureResult {
+	result := input
 	structuredStreams := [][]byte{result.Stderr, result.Stdout}
 	if failure, ok := lastClaudeStructuredFailure(structuredStreams); ok {
 		return failure
@@ -52,12 +88,12 @@ func ParseClaudeProviderFailure(result CommandResult) ProviderFailureResult {
 		return failure
 	}
 	if result.ExitCode == 124 {
-		return ProviderFailureResult{
+		return FailureResult{
 			Reason:  interfaces.WorkFailureTypeTimeout,
 			Message: claudeTimeoutFailureMessage,
 		}
 	}
-	return ProviderFailureResult{
+	return FailureResult{
 		Reason:  interfaces.WorkFailureTypeUnknown,
 		Message: claudeUnknownFailureMessage(streams, result.ExitCode),
 	}
@@ -66,8 +102,8 @@ func ParseClaudeProviderFailure(result CommandResult) ProviderFailureResult {
 // Streams are ordered stderr then stdout, and lines retain their stream order.
 // The final recognized record wins; malformed and unrelated later lines do not
 // displace it. This matches structured-record selection above.
-func lastClaudeStructuredFailure(streams [][]byte) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
+func lastClaudeStructuredFailure(streams [][]byte) (FailureResult, bool) {
+	var last FailureResult
 	var found bool
 	for _, stream := range streams {
 		for len(stream) > 0 {
@@ -89,7 +125,7 @@ func lastClaudeStructuredFailure(streams [][]byte) (ProviderFailureResult, bool)
 			if !recognized {
 				continue
 			}
-			last = ProviderFailureResult{
+			last = FailureResult{
 				Reason:  reason,
 				Message: claudeStructuredFailureMessage(failure.Message, reason),
 			}
@@ -214,7 +250,7 @@ func containsClaudeCredentialSignal(message string) bool {
 			prefix = prefix[boundary+1:]
 		}
 		name := strings.Trim(prefix, "\"'{}[](),")
-		if ClassifyCommandEnvKey(name) == CommandEnvClassificationRedacted {
+		if isRedactedCommandEnvKey(name) {
 			return true
 		}
 		remainder = remainder[separator+1:]
@@ -247,7 +283,7 @@ func isClaudeCredentialField(field string) bool {
 	}
 	identifier := strings.ReplaceAll(field, "-", "_")
 	return containsClaudeSensitiveIdentifierPart(identifier) &&
-		ClassifyCommandEnvKey(identifier) == CommandEnvClassificationRedacted
+		isRedactedCommandEnvKey(identifier)
 }
 
 func containsClaudeCredentialWord(message string) bool {
@@ -265,7 +301,7 @@ func containsClaudeSensitiveIdentifier(message string) bool {
 		identifier := strings.Trim(field, "\"'{}[](),.;:!?=")
 		if strings.Contains(identifier, "_") &&
 			containsClaudeSensitiveIdentifierPart(identifier) &&
-			ClassifyCommandEnvKey(identifier) == CommandEnvClassificationRedacted {
+			isRedactedCommandEnvKey(identifier) {
 			return true
 		}
 	}
@@ -282,8 +318,8 @@ func containsClaudeSensitiveIdentifierPart(identifier string) bool {
 	return false
 }
 
-func lastClaudeTextFailure(streams []string) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
+func lastClaudeTextFailure(streams []string) (FailureResult, bool) {
+	var last FailureResult
 	var found bool
 	for _, stream := range streams {
 		for _, line := range strings.Split(stream, "\n") {
@@ -295,7 +331,7 @@ func lastClaudeTextFailure(streams []string) (ProviderFailureResult, bool) {
 			if !ok {
 				continue
 			}
-			last = ProviderFailureResult{
+			last = FailureResult{
 				Reason:  reason,
 				Message: claudeTextFailureMessage(normalized, reason),
 			}
@@ -471,153 +507,12 @@ func boundUTF8Bytes(message string, limit int) string {
 
 }
 
-// ProviderErrorCorpusEntry is one shared raw provider-failure fixture used by
-// worker unit tests and functional smoke coverage.
-type ProviderErrorCorpusEntry struct {
-	Name                  string                       `json:"name"`
-	Provider              interfaces.ModelProvider     `json:"provider"`
-	RawProviderFamily     string                       `json:"raw_provider_family"`
-	Category              string                       `json:"category"`
-	UpstreamSourceCase    string                       `json:"upstream_source_case"`
-	ExitCode              int                          `json:"exit_code"`
-	Stdout                string                       `json:"stdout"`
-	Stderr                string                       `json:"stderr"`
-	ExpectedType          interfaces.WorkFailureType   `json:"expected_type"`
-	ExpectedFamily        interfaces.WorkFailureFamily `json:"expected_family"`
-	ExpectedMessage       string                       `json:"expected_message,omitempty"`
-	Retryable             bool                         `json:"retryable"`
-	TriggersThrottlePause bool                         `json:"triggers_throttle_pause"`
-	Supported             bool                         `json:"supported"`
-	RejectMessageContains []string                     `json:"reject_message_contains,omitempty"`
-	Notes                 string                       `json:"notes,omitempty"`
-}
-
-// CommandResult renders the raw shared fixture into the provider subprocess
-// contract used by normalization tests and smoke harnesses.
-func (e ProviderErrorCorpusEntry) CommandResult() CommandResult {
-	return CommandResult{
-		ExitCode: e.ExitCode,
-		Stdout:   []byte(e.Stdout),
-		Stderr:   []byte(e.Stderr),
-	}
-}
-
-// RepeatedCommandResults expands one shared failure shape into a fixed number
-// of repeated provider command results for bounded retry and throttle tests.
-func (e ProviderErrorCorpusEntry) RepeatedCommandResults(count int) []CommandResult {
-	results := make([]CommandResult, 0, count)
-	for range count {
-		results = append(results, e.CommandResult())
-	}
-	return results
-}
-
-type providerErrorCorpusFile struct {
-	Entries []ProviderErrorCorpusEntry `json:"entries"`
-}
-
-// ProviderErrorCorpus is the cached shared provider-failure fixture set.
-type ProviderErrorCorpus struct {
-	entriesByName map[string]ProviderErrorCorpusEntry
-	allEntries    []ProviderErrorCorpusEntry
-}
-
-// Entry returns the named shared fixture.
-func (c ProviderErrorCorpus) Entry(name string) (ProviderErrorCorpusEntry, bool) {
-	entry, ok := c.entriesByName[name]
-	return entry, ok
-}
-
-// Entries returns all corpus fixtures in stable order.
-func (c ProviderErrorCorpus) Entries() []ProviderErrorCorpusEntry {
-	return append([]ProviderErrorCorpusEntry(nil), c.allEntries...)
-}
-
-// SupportedEntriesForCategory returns the currently supported fixtures for one
-// normalized provider-failure category.
-func (c ProviderErrorCorpus) SupportedEntriesForCategory(category string) []ProviderErrorCorpusEntry {
-	entries := make([]ProviderErrorCorpusEntry, 0, len(c.allEntries))
-	for _, entry := range c.allEntries {
-		if entry.Supported && entry.Category == category {
-			entries = append(entries, entry)
+func containsAny(haystack string, needles ...string) bool {
+	for _, needle := range needles {
+		if needle != "" && strings.Contains(haystack, needle) {
+			return true
 		}
 	}
-	return entries
+	return false
 }
 
-//go:embed testdata/provider_error_corpus.json
-var providerErrorCorpusJSON []byte
-
-var (
-	providerErrorCorpusOnce sync.Once
-	providerErrorCorpus     ProviderErrorCorpus
-	providerErrorCorpusErr  error
-)
-
-// LoadProviderErrorCorpus returns the shared provider-failure fixture corpus.
-func LoadProviderErrorCorpus() (ProviderErrorCorpus, error) {
-	providerErrorCorpusOnce.Do(func() {
-		providerErrorCorpus, providerErrorCorpusErr = loadProviderErrorCorpus()
-	})
-	return providerErrorCorpus, providerErrorCorpusErr
-}
-
-func loadProviderErrorCorpus() (ProviderErrorCorpus, error) {
-	var raw providerErrorCorpusFile
-	if err := json.Unmarshal(providerErrorCorpusJSON, &raw); err != nil {
-		return ProviderErrorCorpus{}, fmt.Errorf("decode provider error corpus: %w", err)
-	}
-	if len(raw.Entries) == 0 {
-		return ProviderErrorCorpus{}, fmt.Errorf("decode provider error corpus: no entries")
-	}
-
-	entriesByName := make(map[string]ProviderErrorCorpusEntry, len(raw.Entries))
-	for _, entry := range raw.Entries {
-		if err := validateProviderErrorCorpusEntry(entry); err != nil {
-			return ProviderErrorCorpus{}, err
-		}
-		if _, exists := entriesByName[entry.Name]; exists {
-			return ProviderErrorCorpus{}, fmt.Errorf("decode provider error corpus: duplicate entry %q", entry.Name)
-		}
-		entriesByName[entry.Name] = entry
-	}
-
-	return ProviderErrorCorpus{
-		entriesByName: entriesByName,
-		allEntries:    append([]ProviderErrorCorpusEntry(nil), raw.Entries...),
-	}, nil
-}
-
-func validateProviderErrorCorpusEntry(entry ProviderErrorCorpusEntry) error {
-	if entry.Name == "" {
-		return fmt.Errorf("decode provider error corpus: missing entry name")
-	}
-	if entry.Provider == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing provider", entry.Name)
-	}
-	if entry.RawProviderFamily == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing raw provider family", entry.Name)
-	}
-	if entry.Category == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing category", entry.Name)
-	}
-	if entry.UpstreamSourceCase == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing upstream source case", entry.Name)
-	}
-	if entry.ExpectedType == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing expected type", entry.Name)
-	}
-	if entry.ExpectedFamily == "" {
-		return fmt.Errorf("decode provider error corpus: entry %q missing expected family", entry.Name)
-	}
-	if entry.Provider == interfaces.ModelProviderClaude && entry.ExpectedMessage == "" {
-		return fmt.Errorf("decode provider error corpus: Claude entry %q missing expected message", entry.Name)
-	}
-	if entry.ExpectedFamily == interfaces.WorkFailureFamilyThrottle && !entry.TriggersThrottlePause {
-		return fmt.Errorf("decode provider error corpus: entry %q throttle family must trigger throttle pause", entry.Name)
-	}
-	if entry.TriggersThrottlePause && !entry.Retryable {
-		return fmt.Errorf("decode provider error corpus: entry %q throttle pause requires retryable=true", entry.Name)
-	}
-	return nil
-}
